IMP: (09/23/09, AM) Allow bold dots on non-integer lambda multiples.
git-svn-id: https://svn.java.net/svn/electric~svn/trunk@10133 b64ed6c1-75a8-e629-b1da-d0dae74a743a

+++ srcj/com/sun/electric/tool/user/ui/ClickZoomWireListener.java
@@ -1491,8 +1491,8 @@ public class ClickZoomWireListener
 		Dimension2D arrowDistance = User.getAlignmentToGrid();
 		dX *= arrowDistance.getWidth();
 		dY *= arrowDistance.getHeight();
-		int scaleX = User.getDefGridXBoldFrequency();
-        int scaleY = User.getDefGridYBoldFrequency();
+		double scaleX = User.getDefGridXBoldFrequency();
+        double scaleY = User.getDefGridYBoldFrequency();
 		if (scaleMove) { dX *= scaleX;   dY *= scaleY; }
 		if (scaleMove2) { dX *= scaleX;   dY *= scaleY; }
